Fixed some bugs with the integer cell-size for big grids, and also got it so that the 3D blast wave works in parallel and is identical to serial code.  Got silocompare and silo2ascii analysis tools working with new code.

+++ source/sim_control_MPI.cpp
@@ -52,7 +52,7 @@
 ///    added ParallelParams class, and fixing code for non-global mpiPM.
 /// - 2015.02.18 JM: moved setup functions to setup_fixed_grid_MPI
 /// - 2016.03.14 JM: Worked on parallel Grid_v2 update (full
-///    boundaries).
+///    boundaries).  Changed default I/O to DOUBLE precision.
 
 #include "defines/functionality_flags.h"
 #include "defines/testing_flags.h"
@@ -240,7 +240,7 @@ int sim_control_fixedgrid_pllel::Init(
       mpiPM.ReadSingleFile = false;
     }
 
-    if (!dataio) dataio = new dataio_silo_pllel ("FLOAT", &mpiPM);
+    if (!dataio) dataio = new dataio_silo_pllel ("DOUBLE", &mpiPM);
     if (!dataio) rep.error("dataio_silo_pllel initialisation",dataio);
     if (!dataio->file_exists(infile)) {
       cout <<"\tInfile doesn't exist: failing\n";
@@ -299,7 +299,7 @@ int sim_control_fixedgrid_pllel::Init(
 #endif
 #ifdef SILO
     case 5: // silo
-      dataio = new dataio_silo_pllel ("FLOAT", &mpiPM);
+      dataio = new dataio_silo_pllel ("DOUBLE", &mpiPM);
       break;
 #endif // if SILO
     default:
